Separate out a few more sub-modules
