+++ lib/nightmare.js
@@ -24,7 +24,7 @@ const ipc = require('./ipc');
  * runner script
  */
 
-const runner = join(__dirname, 'runner.js');
+const runnerPath = join(__dirname, 'runner.js');
 
 /**
  * Template
@@ -192,7 +192,6 @@ class Nightmare {
                     var results = getFunctionNames(tink, propertyKey);
                     functionNames = functionNames.concat(results);
                 }
-                    
             }
 
             return functionNames;
@@ -322,7 +321,7 @@ class Nightmare {
      */
     init() {
         debug('.init() starting');
-        this.proc = proc.spawn(this._options.electronPath, [runner].concat(JSON.stringify(this._options.electronArgs)), {
+        this.proc = proc.spawn(this._options.electronPath, [runnerPath].concat(JSON.stringify(this._options.electronArgs)), {
             stdio: [null, null, null, 'ipc']
         });
